05-docker/workshop/6/web_task/service/requirements.txt
flask
pymysql
pycrypto